+++ packages/web/app/src/components/v2/heading.tsx
@@ -17,7 +17,7 @@ export const Heading = ({
   return (
     <HeadingLevel
       className={clsx(
-        'text-white',
+        'text-white cursor-default',
         {
           lg: 'text-lg font-bold',
           xl: 'text-xl font-bold',